Feat: move out the merge related code to merge.rs
